# modules/genealogy-media-api/resources/openapi/v1/genealogy-media.yaml
openapi: 3.1.0
info: { title: Genealogy Media API, version: 1.0.0 }
paths:
...
  /api/v1/genealogy/media/{record}/transcribe:
    post:
      operationId: genealogy.media.transcribe
      security: [{ sanctum: [] }]
      responses: { '200': { description: Transcription result } }
...
  /api/v1/genealogy/media/library:
    get:
      operationId: genealogy.media.library
      security: [{ sanctum: [] }]
      parameters: [{ name: q, in: query, schema: { type: string } }, { name: kind, in: query, schema: { type: string } }, { name: public_only, in: query, schema: { type: boolean } }]
      responses: { '200': { description: Privacy-filtered media library results } }
...
components:
  securitySchemes: { sanctum: { type: http, scheme: bearer, bearerFormat: Sanctum } }
...